Added FB messenger module
Ran `botpress install messenger` in /trollbusters.

Got the following warnings:

```
ERR npm WARN botpress-messenger@2.2.9 requires a peer of botpress@>= 0.1 but none was installed.
[botpress] 	 ERR npm WARN react-audio-player@0.5.0 requires a peer of react@15.x but none was installed.
npm WARN react-audio-player@0.5.0 requires a peer of react-dom@15.x but none was installed.
npm WARN react-player@0.18.0 requires a peer of react@* but none was installed.

```

However, it also says:

```
[botpress] 	 added 58 packages in 10.392s
[botpress] 	 OK Modules successfully installed

```

So I'm going to call it a win.

Responds to Issue #1.